requirements.txt
paho-mqtt==1.6.1
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
